refactor(backend): consolidate orchestrator files (#3393)
Consolidate 7 orchestrator files into 2 organized packages:
- enhanced_orchestration/: EnhancedMultiAgentOrchestrator (moved from enhanced_multi_agent_orchestrator.py)
- agents/agent_orchestration/: AgentOrchestrator (moved from agents/agent_orchestrator.py)

Backward-compatibility shims remain for 1-2 release cycles.
